// giant-bpm-fe/src/types/flow.ts
export enum WorkflowNodeKey {
  Approval = "Approval",
  ParallelApproval = "ParallelApproval",
  Condition = "Condition",
  Subflow = "Subflow",
  Form = "Form",
  End = "End",
  Placeholder = "Placeholder",
  DummyNode = "DummyNode",
}
// ...
interface BaseNodeData extends Record<string, unknown> {
  key?: string;
  label?: string;
  description?: string;
  next: string | null;
  parents: string[] | null;
}
// ...
export interface VisibilityRule {
  componentName: string;
  actions: VisibilityAction[];
  condition?: string;
}
// ...
export type ApplicantSource = "selection" | "submitter";

export interface FormNodeData extends BaseNodeData {
  form?: Pick<
    FormDefinition,
    "id" | "name" | "tags" | "description" | "schema" | "revisionId"
  >;
  formId?: string;
  componentRules?: VisibilityRule[];
  applicantSource?: ApplicantSource;
  serialPrefix?: string;
}
export type FormNodeType = Node<FormNodeData, WorkflowNodeKey.Form>;